feat: modal takes full height - 16px off vertical margins
