+++ apps/web/components/modals/UserProfile/contents/index.tsx
@@ -1,0 +1,22 @@
+import SavedGraphs from './saved-graphs'
+import ModulesDone from './modules-done'
+
+// export default function Contents() {
+//   return (
+//     <div className="mt-2 grid grid-cols-3 space-x-6">
+//       <div>
+//         <h3 className="mb-4">Saved Graphs</h3>
+//         <SavedGraphs />
+//       </div>
+//       <h3>Saved Degrees</h3>
+//       <div>
+//         <h3 className="mb-4">Modules Done</h3>
+//         <ModulesDone />
+//       </div>
+//     </div>
+//   )
+// }
+
+export default function Contents() {
+  return <div className="h-full bg-green-100"></div>
+}
